pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "tech-support-rag-api"
version = "0.1.0"
description = "FastAPI-based Haystack RAG service for local llama.cpp deployments"
readme = "README.md"
authors = [{ name = "Greved" }]
license = { text = "Proprietary" }
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.5",
  "uvicorn[standard]>=0.24.0",
  "pydantic-settings>=2.2.1",
  "haystack-ai[dev]>=2.1.0",
  "qdrant-client>=1.9.0",
  "qdrant-haystack>=0.0.1",
  "typer>=0.12.5",
  "python-dotenv>=1.0.1",
  "loguru>=0.7.2",
  "requests>=2.32.3",
  "PyYAML>=6.0.2",
  "google-genai-haystack>=0.2.2",
  "python-docx>=1.1.0",
  "python-multipart>=0.0.9",
  "prometheus-client>=0.21.0",
  "python-json-logger>=3.2.0",
  "opentelemetry-exporter-otlp>=1.30.0",
  "opentelemetry-sdk>=1.30.0",
  "opentelemetry-instrumentation-fastapi>=0.51b0",
  "opentelemetry-instrumentation-requests>=0.51b0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.8",
  "black>=24.10.0",
  "mypy>=1.12.0",
  "pytest>=8.3.3",
  "pre-commit>=3.8.0"
]

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "I", "UP", "B", "SIM"]
exclude = [
  "docs/*",
]

[tool.ruff.lint.isort]
known-first-party = ["app", "ingestion"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app", "ingestion"]
